feat: use peer-dependencies to handle dependency to three.js (#614)
* feat: use peerDependencies to handle dependency to three.js

BREAKING CHANGE: while it will probably not be a problem for most, this changes how the library has to be loaded for some users. Typescript users will get an error about the THREE-option no longer being valid.

* feat: build-tasks and package configuration

 - updated package-lock.json to lockfile version 3
 - include source-files in published bundle
 - fix problems with non-esm builds not including the polyfills
   required for legacy browsers
 - add exports entrypoint specification with conditional exports

* chore: update README

Added ESM usage to installation section and clarified loading order when
used via UMD in the browser.

* chore: add note on umd-usage to example in readme

---------

Co-authored-by: Alex Muramoto <amuramoto@users.noreply.github.com>