feat: add raycasting support (#639)
* chore: reorganize code

 - reorganized the order of methods
 - add visibility `public` as required by eslint
 - add jsdoc from Maps API
 - minor adjustments to eslint-rules
 - remove obsolete test

* feat: add onBeforeDraw callback

Added a new callback that can be overwritten by users to implement updates of their scene before a map-frame is rendered.

* feat: precise positioning and configurable up-axis

Fixes the issues related to imprecise or jittery positions and adds support for selectable up-axis.
The old version of the lat/lng to world-space conversion functions only worked properly when the
origin of the scene was positioned at {lat:0,lng:0}. But that leads to numeric problems (distorted
meshes and jittery motion) due to excessively large values for the world-space x/y coordinates.

The implementation of `latLngToVector3Relative()` that could be used to compute positions relative
to the anchor-point didn't account for the mercator-scale factor, which leads to very imprecise
positions in higher latitudes.

Additionally, we want the user to be able to choose the up-axis they want to use, as there are
good reasons for both the y-up and z-up orientation, depending on the use-case.

In order to handle the coordinate translations properly at all latitudes and for any world-space
orientation, the conversion functions have to be aware of the current anchor-position and upAxis.
For this reason, the conversion function `latLngAltitudeToVector3` needs to be a method of the
overlay-class instead of a utility function. The utility functions in `util.ts` should be
considered internal and no longer be used outside the library.

This commit also changes the way the orientation is specified. While before the raw Float32Arrays
for the CoordinateTransformer are specified to the constructor, we now only use a single option,
`upAxis` that can be specified as 'Y' or 'Z', or a Vector3 to describe the axis to be used as "up".

BREAKING CHANGE: removed, renamed and refactored exported functions from utils.ts, changed shape of options (removed and replaced parameters)

* feat: add raycast() function and example

Adds a new function `ThreeJSOverlayView.raycast()` that allows users to query the 3D scene for objects at a specified screen-space position.

The obvious use-case for this is add implement mouse-interactions, which is demonstrated in the readme and new example.

Fixes #20.