feat: miscellaneous smaller changes and fixes (#640)
* refactor: setAnchor() instead of public property

This changes the anchor from a public property to a `setAnchor()` method, allowing all latLng types as parameter. This is in preparation of the upcoming floating origin feature.

BREAKING CHANGE: `.anchor` is made protected and `.setAnchor()` has to be used instead.

* chore: prevent rollup from clearing the console

* docs: more docs about coordinates, anchor and transforms

* fix: add animationMode option

Previously the ThreeJSOverlay view would always re-render the entire map with every animation frame, even if that isn't required.

The new default value is to only render new frames when requested by either the user (via `requestRedraw()`) or the Map, which should typically be enough in most situations. Users can still use `animationMode: "always"` to get the old behaviour.

* chore: update dependencies

updates all dependencies to the latest major version

 - change type to 'module' in package.json
 - jest and ts-jest: update configuration for new version
 - rollup
   - enabled sourcemaps for all builds
   - rollup explicit babelHelpers config
   - rollup replace abandoned rollup-plugin-terser with @rollup/plugin-terser
   - explicit external three.js dependency
 - examples: configured inline-sourcemaps for all examples

* feat: add new addDefaultLighting option

A new `addDefaultLighting` option (default `true`) to have basic lighting added to the scene automatically.

* fix: disable collectCoverage in jest-config

collectCoverage adds instrumentation to the files being run which interferes with the debugging ability.

For running the tests via npm-task, coverage is enabled seperately.

* fix: reorganize and extend test-suite

* fix: use jest-webgl-canvas-mock instead of headless gl

* fix: fix build-problems with docs and examples