feat: new API version with breaking changes (#656)
* feat: use peer-dependencies to handle dependency to three.js (#614)

* feat: use peerDependencies to handle dependency to three.js

BREAKING CHANGE: while it will probably not be a problem for most, this changes how the library has to be loaded for some users. Typescript users will get an error about the THREE-option no longer being valid.

* feat: build-tasks and package configuration

 - updated package-lock.json to lockfile version 3
 - include source-files in published bundle
 - fix problems with non-esm builds not including the polyfills
   required for legacy browsers
 - add exports entrypoint specification with conditional exports

* chore: update README

Added ESM usage to installation section and clarified loading order when
used via UMD in the browser.

* chore: add note on umd-usage to example in readme

---------

Co-authored-by: Alex Muramoto <amuramoto@users.noreply.github.com>

* chore: add new npm-task to start a webserver alongside rollup (#636)

* feat: add raycasting support (#639)

* chore: reorganize code

 - reorganized the order of methods
 - add visibility `public` as required by eslint
 - add jsdoc from Maps API
 - minor adjustments to eslint-rules
 - remove obsolete test

* feat: add onBeforeDraw callback

Added a new callback that can be overwritten by users to implement updates of their scene before a map-frame is rendered.

* feat: precise positioning and configurable up-axis

Fixes the issues related to imprecise or jittery positions and adds support for selectable up-axis.
The old version of the lat/lng to world-space conversion functions only worked properly when the
origin of the scene was positioned at {lat:0,lng:0}. But that leads to numeric problems (distorted
meshes and jittery motion) due to excessively large values for the world-space x/y coordinates.

The implementation of `latLngToVector3Relative()` that could be used to compute positions relative
to the anchor-point didn't account for the mercator-scale factor, which leads to very imprecise
positions in higher latitudes.

Additionally, we want the user to be able to choose the up-axis they want to use, as there are
good reasons for both the y-up and z-up orientation, depending on the use-case.

In order to handle the coordinate translations properly at all latitudes and for any world-space
orientation, the conversion functions have to be aware of the current anchor-position and upAxis.
For this reason, the conversion function `latLngAltitudeToVector3` needs to be a method of the
overlay-class instead of a utility function. The utility functions in `util.ts` should be
considered internal and no longer be used outside the library.

This commit also changes the way the orientation is specified. While before the raw Float32Arrays
for the CoordinateTransformer are specified to the constructor, we now only use a single option,
`upAxis` that can be specified as 'Y' or 'Z', or a Vector3 to describe the axis to be used as "up".

BREAKING CHANGE: removed, renamed and refactored exported functions from utils.ts, changed shape of options (removed and replaced parameters)

* feat: add raycast() function and example

Adds a new function `ThreeJSOverlayView.raycast()` that allows users to query the 3D scene for objects at a specified screen-space position.

The obvious use-case for this is add implement mouse-interactions, which is demonstrated in the readme and new example.

Fixes #20.

* feat: miscellaneous smaller changes and fixes (#640)

* refactor: setAnchor() instead of public property

This changes the anchor from a public property to a `setAnchor()` method, allowing all latLng types as parameter. This is in preparation of the upcoming floating origin feature.

BREAKING CHANGE: `.anchor` is made protected and `.setAnchor()` has to be used instead.

* chore: prevent rollup from clearing the console

* docs: more docs about coordinates, anchor and transforms

* fix: add animationMode option

Previously the ThreeJSOverlay view would always re-render the entire map with every animation frame, even if that isn't required.

The new default value is to only render new frames when requested by either the user (via `requestRedraw()`) or the Map, which should typically be enough in most situations. Users can still use `animationMode: "always"` to get the old behaviour.

* chore: update dependencies

updates all dependencies to the latest major version

 - change type to 'module' in package.json
 - jest and ts-jest: update configuration for new version
 - rollup
   - enabled sourcemaps for all builds
   - rollup explicit babelHelpers config
   - rollup replace abandoned rollup-plugin-terser with @rollup/plugin-terser
   - explicit external three.js dependency
 - examples: configured inline-sourcemaps for all examples

* feat: add new addDefaultLighting option

A new `addDefaultLighting` option (default `true`) to have basic lighting added to the scene automatically.

* fix: disable collectCoverage in jest-config

collectCoverage adds instrumentation to the files being run which interferes with the debugging ability.

For running the tests via npm-task, coverage is enabled seperately.

* fix: reorganize and extend test-suite

* fix: use jest-webgl-canvas-mock instead of headless gl

* fix: fix build-problems with docs and examples

---------

Co-authored-by: Alex Muramoto <amuramoto@users.noreply.github.com>